1 Module in 1 Module Out - and new Custom Module Templates.
Summary:This allows one module in to always be compiled to one
module out, and works with a custom template that you supply
which can be used to model various module systems.
It replaced any use of wrap_with_fun because it is possible
to implement wrap_with_fun in terms of custom header templates.

Test Plan:

Reviewers:

CC: