Properly inject custom runtime functions for certain backends.
Summary: Eliminates the need for identifier_renaming. Eliminates
some of the hacky bolt on injection of runtime variables for
unnamed globals found at the top level for PHP.

This diff adds the concept of "optionally included runtime stubs"
which will turn into Rehp nodes if the runtime stubs aren't provided.

It also removes any conversion of JS to Rehp, and instead treats JS as a pure string (when in the PHP compilation pipline) - should do the same for JS pipeline next.

Test Plan:

Reviewers:

CC: