Use Reason-native fp library for file paths.
Summary:Updates most usage of file paths to be reason-native's fp
library, and in the process eliminate bugs (it's very difficult to work
with raw path strings without a little help).
Fp is a single file library, and I've namespaced it so it doesn't
conflict with other ones.

Test Plan:

Reviewers:

CC: