New test case for stale hashes.
Summary:This reproduces an issue with stale hashes when the developer
rewrites a module to be identical in terms of functionality - but with
a series of renames in the code. This _should_ be detected by our debug
data hashing, but it isn't. It does not produce any bugs in pure Reason
compiled programs, but with the way we expose readable names cause bugs
when consumers access these modules via the convenience exported named
PHP functions.

I tried a few things to try to correctly hash the debug data to detect
renames, but they did not work.

Test Plan:

Reviewers:

CC: