update hack output now that we no longer try to import js stubs into php
Summary:This is no longer needed with the new macro system and external runtime mandated for php.
The output for the "runtime only" option for php is wrong obviously, but it's not used and I'll clean it
up at some point.
I don't know why the definition of pi changed though! (It's more precise).

Test Plan:

Reviewers:

CC: