Recursion bug in hack output
There is a bug in the Hack output when recursive functions are created inside if statements or loops.

If `--enable vardecl` is passed, the functions are not put into Ref->contents. If it is not passed, then they are set to Ref->contents, but the ref itself is never created. (In both cases, there is a bug that cause an exception at runtime). If `--noinline` is passed, it will fix the bug in a special case where the recursive function was previously being inlined, but in general this bug doesn't appear related to that flag.