Remove topLevelIdentifiers
The sources function was relying on topLevelIdentifiers to find all variables that are declared in a function in order to figure out which variables should be refs. However, it doesn't traverse very deep into function bodies, so it caused bugs in cases where variables that need to be refs are inside if statements or loops.

Instead of using a function like topLevelIdentifiers, we can just reuse the logic for determining when to add ->contents to variable declaration in foldVars. The logic is: if the variable being declared references itself in its declaration (example: recursive functions) or if it has already been used but not declared earlier in the function body (example: mutually recursive functions) then the variable should be stored in a Ref, so the left hand side of the declaration should use ->contents.

When that it determined, the variable can be recorded in "output.refs" and then used by sources to prepend Ref declarations at the top of the function.

Notes:
- this doesn't fix recursion bugs when the vardecl flag is passed. I'm still trying to understand that case.
- there are a couple of prexisting bugs revealed by the changes to Stdlib__arg and Stdlib__genlex