Fix recursion for vardecl
vardecl replaces variable declarations with EBin(Eq, id, rhs) expressions (variable assignments), and then adds new None variable declarations at the top of the function.

This breaks Ref creation. Ref creation relies on detecting when a variable declaration uses itself, or when a variable being declared has already been used. Replacing declaration ASTs with assignment ASTs and adding new declarations at the top of the function prevent that detection from working.

The fix in this commit is to move the vardecl step to php_from_rehp (it's no longer run for JS). output.dec tracks mostly the same variables that vardecl moves to null declarations. It doesn't include Refs, but input vars and other variables that are declared at the top level of functions need to be removed from it (if "var x=3" is at the top of a function, vardecl will detect it and not add "var x=null").

Detecting top level variables is done in a separate pass, similar to the topLevelIdents function that was removed earlier in this stack.

The changes in rehack_tests show that this fixes the recursion bugs in the recursion tests, as well as a couple more bugs in other stdlib files. Unfortunately, the order of declarations is not retained by output.dec's StringMap, so there's a lot of extra noise.