Add Hash Busting control via environment variable.
Summary:
See the comment in `js_of_ocaml.ml`. It is sometimes useful to be able
to force recompilation.

I also think we should add back a rehp version in the file output's hash
footer at some point.
With this diff you have more control though. Set an env var in your
environment to rebuild. If using something like esy + Dune rules, this
wouldn't usually even be an issue because it would purge all the build
artifacts. But if you are building directly into a monorepo and checking
in the artifacts it could be an issue and the manual hash buster gives
repo owners full control of purging artifacts.

Test Plan:

Reviewers:

CC: